Update copy from User -> Person
